balena-os.inc: Don't include security flags by default
Since thud, poky distro file on which balena OS is based, already
includes security_flags.inc. Because of this change, this version throws
a build warning similar to:

WARNING: Duplicate inclusion for /build/../layers/poky/meta/conf/distro/include/security_flags.inc
in
/build/../layers/meta-resin/meta-balena-thud/conf/distro/include/balena-os-yocto-version.inc

This happens because again, we import `poky` and `security_flags` but
since thud, poky includes security_flags by default. In order to avoid
this warning we import it (security_flags) now using an .inc file at the
level of the yocto version meta-balena layer. There is as well a small
wrinkle additional here. We switch the include statement from `require`
to `include` so new layers (like thud) don't have to carry this hack in
the future.

Signed-off-by: Andrei Gherzan <andrei@balena.io>